Functions

GeometricClusterAnalysis.build_matrix — Method

build_matrix(result; indexed_by_r2 = true)

Distance matrix for the graph filtration

indexedbyr2 = true always work

indexedbyr2 = false requires elements of weigths to be non-negative.

indexedbyr2 = false for the sub-level set of the square-root of non-negative power functions : the k-PDTM or the k-PLM (when determinant of matrices are forced to be 1)

GeometricClusterAnalysis.colorize! — Method

colorize!( colors, μ, weights, points, k, signal, centers, Σ)

Fonction auxiliaire qui, étant donnés k centres, calcule les "nouvelles distances tordues" de tous les points de P, à tous les centres On colorie de la couleur du centre le plus proche. La "distance" à un centre est le carré de la norme de Mahalanobis à la moyenne locale "mean" autour du centre + un poids qui dépend d'une variance locale autour du centre auquel on ajoute le log(det(Σ))

On utilise souvent la fonction mahalanobis. mahalanobis(P,c,Σ) calcule le carré de la norme de Mahalanobis (p-c)^T Σ^{-1}(p-c), pour tout point p, ligne de P. C'est bien le carré ; par ailleurs la fonction inverse la matrice Σ ; on peut décider de lui passer l'inverse de la matrice Σ, en ajoutant "inverted = true".

GeometricClusterAnalysis.divergence_poisson — Method

divergence_poisson(x, y)

Bregman divergence associated with the Poisson distribution

GeometricClusterAnalysis.euclidean_sq_distance — Method

euclidean_sq_distance(x, y)

Euclidian sqaured distance

GeometricClusterAnalysis.hierarchical_clustering_lem — Method

matricehauteur : ``(r{i,j}){i,j} r{i,j}$: time$r$when components$i$and$j`` merge

$r_{i,i}$ : birth time of component $i$.

c : number of components

Stop : components whose lifetime is larger than Stop never die

Seuil : centers born after Seuil are removed

It is possible to select Stop and Seuil after running the algorithm with Stop = Inf and Seuil = Inf

For this, we look at the persistence diagram of the components : (x-axis Birth ; y-axis Death)

storeallcolors = TRUE : in the list Couleurs, we store all configurations of colors, for every step.

Thresholding

GeometricClusterAnalysis.mahalanobis — Method

mahalanobis( x, μ, Σ; inverted = false)

Returns the squared Mahalanobis distance of all rows in x and the vector μ = center with respect to Σ = cov. This is (for vector x) defined as

\[D^2 = (x - \mu)' \Sigma^{-1} (x - \mu)\]

x : vector or matrix of data with, say, p columns.

μ : mean vector of the distribution or second data vector of length p or recyclable to that length.

Σ : covariance matrix p x p of the distribution.

inverted : If true, Σ is supposed to contain the inverse of the covariance matrix.

GeometricClusterAnalysis.mutualinfo — Method

This is a copy-paste from Clustering.jl to avoid the dependency

Add something in the docs...

GeometricClusterAnalysis.noisy_three_curves — Method

noisy_three_curves(npoints, nnoise, sigma, d)

nsignal : number of signal points

nnoise : number of additionnal outliers

Signal points are $x = y+z$ with

$y$ uniform on the 3 curves

$z$ normal with mean 0 and covariance matrix $\sigma * I_d$ (with $I_d$ the identity matrix of $R^d$)

d is the dimension of the data and sigma, the standard deviation of the additive Gaussian noise. When $d>2, y_i = 0$ for $i>=2$; with the notation $y=(y_i)_{i=1..d}$

GeometricClusterAnalysis.performance_measurement — Method

performance_measurement

GeometricClusterAnalysis.sample_outliers — Method

sample_outliers(n_outliers, d; L = 1)

GeometricClusterAnalysis.select_parameters — Function

select_parameters(rng, k, alpha, x, Bregman_divergence, maxiter=100)

k est un nombre ou un vecteur contenant les valeurs des differents k

alpha est un nombre ou un vecteur contenant les valeurs des differents alpha

force_decreasing = true force la courbe de risque a etre decroissante en alpha, on utilise les centres optimaux du alpha precedent.

force_decreasing = false, tous les departs sont aléatoires.

GeometricClusterAnalysis.simule_poissond — Method

simule_poissond(n, lambdas, proba)

GeometricClusterAnalysis.subcolorize — Method

subcolorize(points, signal, result, Indices_depart)

Fonction auxiliaire qui, étant donnés le nuage de points, le nombre de points du signal, le résultat de kpdtm ou de kplm et les indices de départ de la méthode de hclust.jl, calcule les "nouvelles distances tordues" de tous les points de P, à tous les centres dont les indices sont dans les indices de départ. On leur associe le centre le plus proche.

GeometricClusterAnalysis.trimmed_bregman_clustering — Method

function trimmed_bregman_clustering(x, k; α = 0, 
divergence_bregman = euclidean_sq_distance, maxiter = 10, nstart = 1)

n : number of points

d : dimension

Input :

x : sample of n points in R^d - matrix of size n $\times$ d

alpha : proportion of eluted points, because considered as outliers. They are given the label 0

k : number of centers

divergence_bregman : function of two numbers or vectors named x and y, which reviews their Bregman divergence.

maxiter: maximum number of iterations allowed.

nstart: if centers is a number, it is the number of different initializations of the algorithm. Only the best result is kept.

Output :

centers: matrix of size dxk whose columns represent the centers of the clusters

cluster: vector of integers in 1:k indicating the index of the cluster to which each point (line) of x is associated.

risk: average of the divergences of the points of x at their associated center.

divergence: the vector of divergences of the points of x at their nearest center in centers, for divergence_bregman.