Trimmed Bregman Clustering

Les divergences de Bregman

Définition de base

Les divergences de Bregman sont des mesures de différence entre deux points. Elles dépendent d'une fonction convexe. Le carré de la distance Euclidienne est une divergence de Bregman. Les divergences de Bregman ont été introduites par Bregman L. M. Bregman (1967).

Soit $\phi$, une fonction strictement convexe et $\mathcal{C}^1$ à valeurs réelles, définie sur un sous ensemble convexe $\Omega$ de $\mathcal{R}^d$. La divergence de Bregman associée à la fonction $\phi$ est la fonction $\mathrm{d}_\phi$ définie sur $\Omega\times\Omega$ par : $\forall x,y\in\Omega,\,{\rm d\it}_\phi(x,y) = \phi(x) - \phi(y) - \langle\nabla\phi(y),x-y\rangle.$

La divergence de Bregman associée au carré de la norme Euclidienne, $\phi:x\in\mathcal{R}^d\mapsto\|x\|^2\in\mathcal{R}$ est égale au carré de la distance Euclidienne :

\[\forall x,y\in\mathcal{R}^d, {\rm d\it}_\phi(x,y) = \|x-y\|^2.\]

Soit $x,y\in\mathcal{R}^d$,

\[\begin{aligned} {\rm d\it}_\phi(x,y) & = \phi(x) - \phi(y) - \langle\nabla\phi(y),x-y\rangle \\ & = \|x\|^2 - \|y\|^2 - \langle 2y, x-y\rangle \\ & = \|x\|^2 - \|y\|^2 - 2\langle y, x\rangle + 2\|y\|^2 \\ & = \|x-y\|^2. \end{aligned}\]

Le lien avec certaines familles de lois

Pour certaines distributions de probabilité définies sur $\mathcal{R}$, d'espérance $\mu\in\mathcal{R}$, la densité ou la fonction de probabilité (pour les variables discrètes), $x\mapsto p_{\phi,\mu,f}(x)$, s'exprime en fonction d'une divergence de Bregman A. Banerjee, S. Merugu, I.S. Dhillon, J. Ghosh (2005) entre $x$ et l'espérance $\mu$ :

\[\begin{equation} p_{\phi,\mu,f}(x) = \exp(-\mathrm{d}_\phi(x,\mu))f(x). \label{eq:familleBregman} \end{equation}\]

Ici, $\phi$ est une fonction strictement convexe et $f$ est une fonction positive.

Certaines distributions sur $\mathcal{R}^d$ satisfont cette même propriété. C'est en particulier le cas des distributions de vecteurs aléatoires dont les coordonnées sont des variables aléatoires indépendantes de lois sur $\mathcal{R}$ du type \eqref(eq:familleBregman).

Soit $Y = (X_1,X_2,\ldots,X_d)$, un $d$-échantillon de variables aléatoires indépendantes, de lois respectives $p_{\phi_1,\mu_1,f_1},p_{\phi_2,\mu_2,f_2},\ldots, p_{\phi_d,\mu_d,f_d}$.

Alors, la loi de $Y$ est aussi du type \eqref{eq:familleBregman}.

La fonction convexe associée est

\[(x_1,x_2,\ldots, x_d)\mapsto\sum_{i = 1}^d\phi_i(x_i).\]

La divergence de Bregman est définie par :

\[((x_1,x_2,\ldots,x_d),(\mu_1,\mu_2,\ldots,\mu_d))\mapsto\sum_{i = 1}^d\mathrm{d}_{\phi_i}(x_i,\mu_i).\]

Soit $X_1,X_2,\ldots,X_d$ des variables aléatoires telles que décrites dans le théorème. Ces variables sont indépendantes, donc la densité ou la fonction de probabilité en $(x_1,x_2,\ldots, x_d)\in\mathcal{R}^d$ est donnée par :

\[\begin{align*} p(x_1,x_2,\ldots, x_d) & = \prod_{i = 1}^dp_{\phi_i,\mu_i,f_i}(x_i)\\ & = \exp\left(-\sum_{i = 1}^d\mathrm{d}_{\phi_i}(x_i,\mu_i)\right)\prod_{i = 1}^df_i(x_i). \end{align*}\]

Par ailleurs, $((x_1,x_2,\ldots,x_d),(\mu_1,\mu_2,\ldots,\mu_d))\mapsto\sum_{i = 1}^d\mathrm{d}_{\phi_i}(x_i,\mu_i)$ est bien la divergence de Bregman associée à la fonction

\[\tilde\phi: (x_1,x_2,\ldots, x_d)\mapsto\sum_{i = 1}^d\phi_i(x_i).\]

En effet, puisque $\nabla\tilde\phi(y_1,y_2,\ldots, y_d) = (\phi_1'(y_1),\phi_2'(y_2),\ldots,\phi_d'(y_d))^T,$ la divergence de Bregman associée à $\tilde\phi$s'écrit :

\[\begin{align*} \tilde\phi & (x_1,x_2,\ldots, x_d) - \tilde\phi(y_1,y_2,\ldots, y_d) - \langle\nabla\tilde\phi(y_1,y_2,\ldots, y_d), (x_1-y_1,x_2-y_2,\ldots, x_d-y_d)^T\rangle\\ & = \sum_{i = 1}^d \left(\phi_i(x_i) - \phi_i(y_i) - \phi_i'(y_i)(x_i-y_i)\right)\\ & = \sum_{i = 1}^d\mathrm{d}_{\phi_i}(x_i,y_i). \end{align*}\]

La divergence associée à la loi de Poisson

La loi de Poisson est une distribution de probabilité sur $\mathcal{R}$ du type \eqref{eq:familleBregman}.

Soit $\mathcal{P}(\lambda)$ la loi de Poisson de paramètre $\lambda>0$. Soit $p_\lambda$ sa fonction de probabilité.

Cette fonction est du type \eqref{eq:familleBregman} pour la fonction convexe

\[\phi: x\in\mathcal{R}_+^*\mapsto x\ln(x)\in\mathcal{R}.\]

La divergence de Bregman associée, $\mathrm{d}_{\phi}$, est définie pour tous $x,y\in\mathcal{R}_+^*$ par :

\[\mathrm{d}_{\phi}(x,y) = x\ln\left(\frac{x}{y}\right) - (x-y).\]

Soit $\phi: x\in\mathcal{R}_+^*\mapsto x\ln(x)\in\mathcal{R}$. La fonction $\phi$ est strictement convexe, et la divergence de Bregman associée à $\phi$ est définie pour tous $x,y\in\mathcal{R}_+$ par :

\[\begin{align*} \mathrm{d}_{\phi}(x,y) & = \phi(x) - \phi(y) - \phi'(y)\left(x-y\right)\\ & = x\ln(x) - y\ln(y) - (\ln(y) + 1)\left(x-y\right)\\ & = x\ln\left(\frac{x}{y}\right) - (x-y). \end{align*}\]

Par ailleurs,

\[\begin{align*} p_\lambda(x) & = \frac{\lambda^x}{x!}\exp(-\lambda)\\ & = \exp\left(x\ln(\lambda) - \lambda\right)\frac{1}{x!}\\ & = \exp\left(-\left(x\ln\left(\frac x\lambda\right) - (x-\lambda)\right) + x\ln(x) - x\right)\frac{1}{x!}\\ & = \exp\left(-\mathrm{d}_\phi(x,\lambda)\right)f(x), \end{align*}\]

avec

\[f(x) = \frac{\exp(x\left(\ln(x) - 1\right))}{x!}.\]

Le paramètre $\lambda$ correspond bien à l'espérance de la variable $X$ de loi $\mathcal{P}(\lambda)$.

Ainsi, d'après le Théorème \@ref(thm:loiBregmanmultidim), la divergence de Bregman associée à la loi d'un $d$-échantillon $(X_1,X_2,\ldots,X_d)$ de $d$ variables aléatoires indépendantes de lois de Poisson de paramètres respectifs $\lambda_1,\lambda_2,\ldots,\lambda_d$ est :

\[\begin{equation} \mathrm{d}_\phi((x_1,x_2,\ldots,x_d),(y_1,y_2,\ldots,y_d)) = \sum_{i = 1}^d \left(x_i\ln\left(\frac{x_i}{y_i}\right) - (x_i-y_i)\right). \label{eq:divBregmanPoisson} \end{equation}\]

Partitionner des données à l'aide de divergences de Bregman

Soit $\mathbb{X} = \{X_1, X_2,\ldots, X_n\}$ un échantillon de $n$ points dans $\mathcal{R}^d$.

Partitionner $\mathbb{X}$ en $k$ groupes revient à associer une étiquette dans $[\![1,k]\!]$ à chacun des $n$ points. La méthode de partitionnement avec une divergence de Bregman A. Banerjee, S. Merugu, I.S. Dhillon, J. Ghosh (2005) consiste en fait à associer à chaque point un centre dans un dictionnaire $\mathbf{c} = (c_1, c_2,\ldots c_k)\in\mathcal{R}^{d\times k}$. Pour chaque point, le choix sera fait de sorte à minimiser la divergence au centre.

Le dictionnaire $\mathbf{c} = (c_1, c_2,\ldots c_k)$ choisi est celui qui minimise le risque empirique

\[R_n:((c_1, c_2,\ldots c_k),\mathbb{X})\mapsto\frac1n\sum_{i = 1}^n\gamma_\phi(X_i,\mathbf{c}) = \frac1n\sum_{i = 1}^n\min_{l\in[\![1,k]\!]}\mathrm{d}_\phi(X_i,c_l).\]

Lorsque $\phi = \|\cdot\|^2$, $R_n$ est le risque associé à la méthode de partitionnement des $k$-means S.P. Lloyd (1982).

L'élagage ou le "Trimming"

Dans J.A. Cuesta-Albertos, A. Gordaliza, C. Matràn (1997), Cuesta-Albertos et al. ont défini et étudié une version élaguée du critère des $k$-means. Cette version permet de se débarrasser d'une certaine proportion $\alpha$ des données, celles que l'on considère comme des données aberrantes. Nous pouvons facilement généraliser cette version élaguée aux divergences de Bregman.

Pour $\alpha\in[0,1]$, et $a = \lfloor\alpha n\rfloor$, la partie entière inférieure de $\alpha n$, la version $\alpha$-élaguée du risque empirique est définie par :

\[R_{n,\alpha}:(\mathbf{c},\mathbb{X})\in\mathcal{R}^{d\times k}\times\mathcal{R}^{d\times n}\mapsto\inf_{\mathbb{X}_\alpha\subset \mathbb{X}, |\mathbb{X}_\alpha| = n-a}R_n(\mathbf{c},\mathbb{X}_\alpha).\]

Ici, $|\mathbb{X}_\alpha|$ représente le cardinal de $\mathbb{X}_\alpha$.

Minimiser le risque élagué $R_{n,\alpha}(\cdot,\mathbb{X})$ revient à sélectionner le sous-ensemble de $\mathbb{X}$ de $n-a$ points pour lequel le critère empirique optimal est le plus faible. Cela revient à choisir le sous-ensemble de $n-a$ points des données qui peut être le mieux résumé par un dictionnaire de $k$ centres, pour la divergence de Bregman $\mathrm{d}_\phi$.

On note $\hat{\mathbf{c}}_{\alpha}$ un minimiseur de $R_{n,\alpha}(\cdot,\mathbb{X})$.

Implémentation de la méthode de partitionnement élagué des données, avec des divergences de Bregman

L'algorithme de partitionnement sans élagage

L'algorithme de S.P. Lloyd (1982) consiste à chercher un minimum $\hat{\mathbf{c}}$ local du risque $R_n(\cdot,\mathbb{X})$ pour le critère des $k$-means (c'est-à-dire, lorsque $\phi = \|\cdot\|^2$). Il s'adapte aux divergences de Bregman quelconques. Voici le fonctionnement de l'algorithme.

Après avoir initialisé un ensemble de $k$ centres $\mathbf{c}_0$, nous alternons deux étapes. Lors de la $t$-ième itération, nous partons d'un dictionnaire $\mathbf{c}_t$ que nous mettons à jour de la façon suivante :

Décomposition de l'échantillon $\mathbb{X}$ selon les cellules de Bregman-Voronoï de $\mathbf{c}_t$ : On associe à chaque point $x$ de l'échantillon $\mathbb{X}$, son centre $c\in\mathbf{c}_t$ le plus proche, i.e., tel que $\mathrm{d}_\phi(x,c)$ soit le plus faible. On obtient ainsi $k$ cellules, chacune associée à un centre ;

Mise à jour des centres : On remplace les centres du dictionnaire $\mathbf{c}_t$ par les barycentres des points des cellules, ce qui donne un nouveau dictionnaire : $\mathbf{c}_{t+1}$.

Une telle procédure assure la décroissance de la suite $(R_n(\mathbf{c}_t,\mathbb{X}))_{t\in\mathcal{N}}$.

Soit $(\mathbf{c}_t)_{t\in\mathcal{N}}$, la suite définie ci-dessus. Alors, pour tout $t\in\mathcal{N}$,

\[R_n(\mathbf{c}_{t+1},\mathbb{X})\leq R_n(\mathbf{c}_t,\mathbb{X}).\]

D'après A. Banerjee, X. Guo, H. Wang (2005), pour toute divergence de Bregman $\mathrm{d}_\phi$ et tout ensemble de points $\mathbb{Y} = \{Y_1,Y_2,\ldots,Y_q\}$, $\sum_{i = 1}^q\mathrm{d}_\phi(Y_i,c)$ est minimale en $c = \frac{1}{q}\sum_{i = 1}^qY_i$.

Soit $l\in[\![1,k]\!]$ et $t\in\mathcal{N}$, notons $\mathcal{C}_{t,l} = \{x\in\mathbb{X}\mid \mathrm{d}_\phi(x,c_{t,l}) = \min_{l'\in [\![1,k]\!]}\mathrm{d}_\phi(x,c_{t,l'})\}$.

Posons $c_{t+1,l} = \frac{1}{|\mathcal{C}_{t,l}|}\sum_{x\in\mathcal{C}_{t,l}}x$. Avec ces notations,

\[\begin{align*} R_n(\mathbf{c}_{t+1},\mathbb{X}) & = \frac1n\sum_{i = 1}^n\min_{l\in[\![1,k]\!]}\mathrm{d}_\phi(X_i,c_{t+1,l})\\ &\leq \frac1n\sum_{l = 1}^{k}\sum_{x\in\mathcal{C}_{t,l}}\mathrm{d}_\phi(x,c_{t+1,l})\\ &\leq \frac1n\sum_{l = 1}^{k}\sum_{x\in\mathcal{C}_{t,l}}\mathrm{d}_\phi(x,c_{t,l})\\ & = R_n(\mathbf{c}_{t},\mathbb{X}). \end{align*}\]

L'algorithme de partitionnement avec élagage

Il est aussi possible d'adapter l'algorithme élagué des $k$-means de J.A. Cuesta-Albertos, A. Gordaliza, C. Matràn (1997). Nous décrivons ainsi cet algorithme, permettant d'obtenir un minimum local du critère $R_{n,\alpha}(.,\mathbb{X})$ :

$\qquad$ INPUT: $\mathbb{X}$ un nuage de $n$ points ; $k\in[\![1,n]\!]$ ; $a\in[\![0,n-1]\!]$ ;

$\qquad$ Tirer uniformément et sans remise $c_1$, $c_2$, $\ldots$, $c_k$ de $\mathbb{X}$.

$\qquad$ WHILE les $c_i$ varient :

$\qquad\qquad$ FOR $i$ dans $[\![1,k]\!]$ :

$\qquad\qquad\qquad$ Poser $\mathcal{C}(c_i)=\{\}$ ;

$\qquad\qquad$ FOR $j$ dans $[\![1,n]\!]$ :

$\qquad\qquad\qquad$ Ajouter $X_j$ à la cellule $\mathcal{C}(c_i)$ telle que $\forall l\neq i,\,\mathrm{d}_{\phi}(X_j,c_i)\leq\mathrm{d}_\phi(X_j,c_l)\,$ ;

$\qquad\qquad\qquad$ Poser $c(X) = c_i$ ;

$\qquad\qquad$ Trier $(\gamma_\phi(X) = \mathrm{d}_\phi(X,c(X)))$ pour $X\in \mathbb{X}$ ;

$\qquad\qquad$ Enlever les $a$ points $X$ associés aux $a$ plus grandes valeurs de $\gamma_\phi(X)$, de leur cellule $\mathcal{C}(c(X))$ ;

$\qquad$ FOR $i$ dans $[\![1,k]\!]$ :

$\qquad\qquad$ $c_i={{1}\over{|\mathcal{C}(c_i)|}}\sum_{X\in\mathcal{C}(c_i)}X$ ;

$\qquad$ OUTPUT: $(c_1,c_2,\ldots,c_k)$;

Ce code permet de calculer un minimum local du risque élagué $R_{n,\alpha = \frac{a}{n}}(\cdot,\mathbb{X})$.

En pratique, il faut ajouter quelques lignes dans le code pour :

traiter le cas où des cellules se vident,

recalculer les étiquettes des points et leur risque associé, à partir des centres $(c_1,c_2,\ldots,c_k)$ en sortie d'algorithme,

proposer la possibilité de plusieurs initialisations aléatoires et retourner le dictionnaire pour lequel le risque est minimal,

limiter le nombre d'itérations de la boucle WHILE,

proposer en entrée de l'algorithme un dictionnaire $\mathbf{c}$, à la place de $k$, pour une initialisation non aléatoire,

éventuellement paralléliser...

L'implémentation

Quelques divergences de Bregman

La fonction poisson calcule la divergence de Bregman associée à la loi de Poisson entre xet y en dimension $d\in^*$. \eqref(eq:divBregmanPoisson)

La fonction euclidean calcule le carré de la norme Euclidienne entre x et y en dimension $d\in\mathcal{N}^*$.

Le code pour le partitionnement élagué avec divergence de Bregman

La méthode de partitionnement élagué avec une divergence de Bregman est codée dans la fonction suivante, trimmed_bregman_clustering, dont les arguments sont :

x : une matrice de taille $n\times d$ représentant les coordonnées des $n$ points de dimension $d$ à partitionner,

centers : un ensemble de centres ou un nombre $k$ correspondant au nombre de groupes,

alpha : dans $[0,1[$, la proportion de points de l'échantillon à retirer ; par défaut 0 (pas d'élagage),

divergence_bregman : la divergence à utiliser ; par défaut euclidean, le carré de la norme Euclidienne (on retrouve le k-means élagué de J.A. Cuesta-Albertos, A. Gordaliza, C. Matràn (1997), tkmeans),

maxiter : le nombre maximal d'itérations,

nstart : le nombre d'initialisations différentes de l'algorithme (on garde le meilleur résultat).

La sortie de cette fonction est une liste dont les arguments sont :

centers : matrice de taille $d\times k$ dont les $k$ colonnes représentent les $k$ centres des groupes,

cluster : vecteur d'entiers dans $[\![0,k]\!]$ indiquant l'indice du groupe auquel chaque point (chaque ligne) de x est associé, l'étiquette $0$ est assignée aux points considérés comme des données aberrantes,

risk : moyenne des divergences des points de x (non considérés comme des données aberrantes) à leur centre associé,

divergence : le vecteur des divergences des points de x à leur centre le plus proche dans centers, pour la divergence divergence_bregman.

GeometricClusterAnalysis.trimmed_bregman_clustering — Function

function trimmed_bregman_clustering(x, k; α = 0, 
divergence_bregman = euclidean_sq_distance, maxiter = 10, nstart = 1)

n : number of points

d : dimension

Input :

x : sample of n points in R^d - matrix of size n $\times$ d

alpha : proportion of eluted points, because considered as outliers. They are given the label 0

k : number of centers

divergence_bregman : function of two numbers or vectors named x and y, which reviews their Bregman divergence.

maxiter: maximum number of iterations allowed.

nstart: if centers is a number, it is the number of different initializations of the algorithm. Only the best result is kept.

Output :

centers: matrix of size dxk whose columns represent the centers of the clusters

cluster: vector of integers in 1:k indicating the index of the cluster to which each point (line) of x is associated.

risk: average of the divergences of the points of x at their associated center.

divergence: the vector of divergences of the points of x at their nearest center in centers, for divergence_bregman.

function trimmed_bregman_clustering(x, centers, α, bregman, maxiter)

n : number of points

d : dimension

Input :

x : sample of n points in R^d - matrix of size n $\times$ d

centers : intial centers

alpha : proportion of eluted points, because considered as outliers. They are given the label 0

bregman : function of two numbers or vectors named x and y, which reviews their Bregman divergence.

maxiter: maximum number of iterations allowed.

Output :

centers: matrix of size dxk whose columns represent the centers of the clusters

risk: average of the divergences of the points of x at their associated center.

Sélection des paramètres $k$ et $\alpha$

Le paramètre $\alpha\in[0,1)$ représente la proportion de points des données à retirer. Nous considérons que ce sont des données aberrantes et leur attribuons l'étiquette $0$.

Afin de sélectionner le meilleur paramètre $\alpha$, il suffit, pour une famille de paramètres $\alpha$, de calculer le coût optimal $R_{n,\alpha}(\hat{\mathbf{c}}_\alpha)$ obtenu à partir du minimiseur local $\hat{\mathbf{c}}_\alpha$ de $R_{n,\alpha}$ en sortie de l'algorithme trimmed_bregman_clustering.

Nous représentons ensuite $R_{n,\alpha}(\hat{\mathbf{c}}_\alpha)$ en fonction de $\alpha$ sur un graphique. Nous pouvons représenter de telles courbes pour différents nombres de groupes, $k$. Une heuristique permettra de choisir les meilleurs paramètres $k$ et $\alpha$.

La fonction GeometricClusterAnalysis.select_parameters, parallélisée, permet de calculer le critère optimal $R_{n,\alpha}(\hat{\mathbf{c}}_\alpha)$ pour différentes valeurs de $k$ et de $\alpha$, sur les données x.

GeometricClusterAnalysis.select_parameters — Function

select_parameters(rng, k, alpha, x, bregman, maxiter=100)

On utilise des centres initiaux aléatoire

k est un nombre ou un vecteur contenant les valeurs des differents k

alpha est un nombre ou un vecteur contenant les valeurs des differents alpha

Mise en œuvre de l'algorithme

Nous étudions les performances de notre méthode de partitionnement de données élagué, avec divergence de Bregman, sur différents jeux de données. En particulier, nous comparons l'utilisation du carré de la norme Euclidienne et de la divergence de Bregman associée à la loi de Poisson. Rappelons que notre méthode avec le carré de la norme Euclidienne coïncide avec la méthode de "trimmed $k$-means" J.A. Cuesta-Albertos, A. Gordaliza, C. Matràn (1997).

Nous appliquons notre méthode à trois types de jeux de données :

Un mélange de trois lois de Poisson en dimension 1, de paramètres $\lambda\in\{10,20,40\}$, corrompues par des points générés uniformément sur $[0,120]$ ;

Un mélange de trois lois de Poisson en dimension 2 (c'est-à-dire, la loi d'un couple de deux variables aléatoires indépendantes de loi de Poisson), de paramètres $(\lambda_1,\lambda_2)\in\{(10,10),(20,20),(40,40)\}$, corrompues par des points générés uniformément sur $[0,120]\times[0,120]$ ;

Les données des textes d'auteurs.

Les poids devant chaque composante des mélanges des lois de Poisson sont $\frac13$, $\frac13$, $\frac13$. Ce qui signifie que chaque variable aléatoire a une chance sur trois d'avoir été générée selon chacune des trois lois de Poisson.

Nous allons donc comparer l'utilisation de la divergence de Bregman associée à la loi de Poisson à celle du carré de la norme Euclidienne, en particulier à l'aide de l'information mutuelle normalisée (NMI). Nous allons également appliquer une heuristique permettant de choisir les paramètres k (nombre de groupes) et alpha (proportion de données aberrantes) à partir d'un jeu de données.

Données de loi de Poisson en dimension 1

Simulation des variables selon un mélange de lois de Poisson

La fonction sample_poisson permet de simuler des variables aléatoires selon un mélange de $k$ lois de Poisson en dimension $d$, de paramètres donnés par la matrice lambdas de taille $k\times d$. Les probabilités associées à chaque composante du mélange sont données dans le vecteur proba.

La fonction sample_outliers permet de simuler des variables aléatoires uniformément sur l'hypercube $[0,L]^d$. On utilisera cette fonction pour générer des données aberrantes.

GeometricClusterAnalysis.sample_poisson — Function

sample_poisson(rng, n, d, lambdas, proba)

GeometricClusterAnalysis.sample_outliers — Function

sample_outliers(rng, n_outliers, d; scale = 1)

On génère un premier échantillon de 950 points de loi de Poisson de paramètre $10$, $20$ ou $40$ avec probabilité $\frac13$, puis un échantillon de 50 données aberrantes de loi uniforme sur $[0,120]$. On note x l'échantillon ainsi obtenu.

using GeometricClusterAnalysis
import GeometricClusterAnalysis: sample_poisson, sample_outliers, performance
using Plots
using Random

n = 1000
n_outliers = 50
d = 1

rng = MersenneTwister(1)
lambdas =  [10,20,40]
proba = [1/3,1/3,1/3]
points, labels = sample_poisson(rng, n - n_outliers, d, lambdas, proba)

outliers = sample_outliers(rng, n_outliers, 1; scale = 120)

x = hcat(points, outliers)
labels_true = vcat(labels, zeros(Int, n_outliers))
scatter( x[1,:], c = labels_true, palette = :rainbow)

Partitionnement des données sur un exemple

Pour partitionner les données, nous utiliserons les paramètres suivants.

k = 3 # Nombre de groupes dans le partitionnement
alpha = 0.04 # Proportion de donnees aberrantes
maxiter = 50 # Nombre maximal d'iterations
nstart = 20 # Nombre de departs

20

Application de l'algorithme classique de $k$-means élagué J.A. Cuesta-Albertos, A. Gordaliza, C. Matràn (1997)

Dans un premier temps, nous utilisons notre algorithme trimmed_bregman_clustering avec le carré de la norme Euclidienne euclidean.

tb_kmeans = trimmed_bregman_clustering(rng, x, k, alpha, euclidean, maxiter, nstart)
tb_kmeans.centers

1×3 Matrix{Float64}:
 41.418  22.5122  10.3412

Nous avons effectué un simple algorithme de $k$-means élagué, comme J.A. Cuesta-Albertos, A. Gordaliza, C. Matràn (1997). On voit trois groupes de même diamètre. Ce qui fait que le groupe centré en $10$ contient aussi des points du groupe centré en $20$. En particulier, les estimations tB_kmeans.centers des moyennes par les centres ne sont pas très bonnes. Les deux moyennes les plus faibles sont bien supérieures aux vraies moyennes $10$ et $20$.

plot(tb_kmeans)

Choix de la divergence de Bregman associée à la loi de Poisson

Lorsque l'on utilise la divergence de Bregman associée à la loi de Poisson, les groupes sont de diamètres variables et sont particulièrement adaptés aux données. En particulier, les estimations tB_Poisson$centers des moyennes par les centres sont bien meilleures.

tb_poisson = trimmed_bregman_clustering(rng, x, k, alpha, poisson, maxiter, nstart)
tb_poisson.centers

1×3 Matrix{Float64}:
 40.8243  10.4911  21.6265

plot(tb_poisson)

Comparaison des performances

Nous mesurons directement la performance des deux partitionnements (avec le carré de la norme Euclidienne, et avec la divergence de Bregman associée à la loi de Poisson), à l'aide de l'information mutuelle normalisée.

Pour le k-means elague :

mutualinfo(labels_true,tb_kmeans.cluster, true)

0.6416317751738038

Pour le partitionnement elague avec divergence de Bregman associee a la loi de Poisson :

mutualinfo(labels_true,tb_poisson.cluster, true)

0.6720639083888875

L'information mutuelle normalisée est supérieure pour la divergence de Bregman associée à la loi de Poisson. Ceci illustre le fait que sur cet exemple, l'utilisation de la bonne divergence permet d'améliorer le partitionnement, par rapport à un $k$-means élagué basique.

Mesure de la performance

Afin de s'assurer que la méthode avec la bonne divergence de Bregman est la plus performante, nous répétons l'expérience précédente replications fois.

Pour ce faire, nous appliquons l'algorithme trimmed_bregman_clustering, sur replications échantillons de taille $n = 1000$, sur des données générées selon la même procédure que l'exemple précédent.

La fonction performance_measurement permet de le faire.

sample_generator = (rng, n) -> sample_poisson(rng, n, d, lambdas, proba)
outliers_generator = (rng, n) -> sample_outliers(rng, n, d; scale = 120)

#3 (generic function with 1 method)

Valeurs par défault: maxiter = 100, nstart = 10, replications = 100

n = 1200
n_outliers = 200
k = 3
alpha = 0.1
nmi_kmeans, _, _ = performance(n, n_outliers, k, alpha, sample_generator, outliers_generator, euclidean)
nmi_poisson, _, _ = performance(n, n_outliers, k, alpha, sample_generator, outliers_generator, poisson)

([0.5062363706331514, 0.5763910882434974, 0.5704150027113549, 0.6379203957624394, 0.6200543255420755, 0.5086997589516993, 0.6269395768318125, 0.6302266736850226, 0.6301054690512806, 0.6362326725904385  …  0.6395263405175107, 0.6159953287841311, 0.5888934822224912, 0.6311777089409022, 0.6039387785354113, 0.6267631247808866, 0.5825978696858187, 0.5993349158906023, 0.6338923119186022, 0.6171748812742635], 0.5976040509922059, 0.010825264297288006)

Les boîtes à moustaches permettent de se faire une idée de la répartition des NMI pour les deux méthodes différentes. On voit que la méthode utilisant la divergence de Bregman associée à la loi de Poisson est la plus performante.

using StatsPlots

boxplot( ones(100), nmi_kmeans, label = "kmeans" )
boxplot!( fill(2, 100), nmi_poisson, label = "poisson" )

Sélection des paramètres $k$ et $\alpha$

On garde le même jeu de données x.

vect_k = collect(1:5)
vect_alpha = sort([((0:2)./50)...,((1:4)./5)...])

params_risks = select_parameters_nonincreasing(rng, vect_k, vect_alpha, x, poisson, maxiter, nstart)

plot(; title = "select parameters")
for (i,k) in enumerate(vect_k)
   plot!( vect_alpha, params_risks[i, :], label ="k=$k", markershape = :circle )
end
xlabel!("alpha")
ylabel!("NMI")

D'après la courbe, on voit qu'on gagne beaucoup à passer de 1 à 2 groupes, puis à passer de 2 à 3 groupes. Par contre, on gagne très peu, en termes de risque, à passer de 3 à 4 groupes ou à passer de 4 à 5 groupes, car les courbes associées aux paramètres $k = 3$, $k = 4$ et $k = 5$ sont très proches. Ainsi, on choisit de partitionner les données en $k = 3$ groupes.

La courbe associée au paramètre $k = 3$ diminue fortement puis à une pente qui se stabilise aux alentours de $\alpha = 0.04$.

Pour plus de précisions concernant le choix du paramètre $\alpha$, nous pouvons nous concentrer sur la courbe $k = 3$ en augmentant la valeur de nstart et en nous concentrant sur les petites valeurs de $\alpha$.

vect_k = [3]
vec_alpha = collect(0:15) ./ 200
params_risks = select_parameters_nonincreasing(rng, [3], vec_alpha, x, poisson, maxiter, 5)

plot(vec_alpha, params_risks[1, :], markershape = :circle)

On ne voit pas de changement radical de pente mais on voit que la pente se stabilise après $\alpha = 0.03$. Nous choisissons le paramètre $\alpha = 0.03$.

Voici finalement le partitionnement obtenu après sélection des paramètres k et alpha selon l'heuristique.

k, alpha = 3, 0.03
tb_poisson = trimmed_bregman_clustering( rng, x, k, alpha, poisson, maxiter, nstart )
tb_poisson.centers

1×3 Matrix{Float64}:
 20.0203  41.1271  9.15363

plot( tb_poisson )

Données de loi de Poisson en dimension 2

Simulation des variables selon un mélange de lois de Poisson

On génère un second échantillon de 950 points dans $\mathcal{R}^2$. Les deux coordonnées de chaque point sont indépendantes, générées avec probabilité $\frac13$ selon une loi de Poisson de paramètre $10$, $20$ ou bien $40$. Puis un échantillon de 50 données aberrantes de loi uniforme sur $[0,120]\times[0,120]$ est ajouté à l'échantillon. On note x l’échantillon ainsi obtenu.

using GeometricClusterAnalysis
import GeometricClusterAnalysis: sample_poisson, sample_outliers, performance
using Plots
using Random

n = 1000
n_outliers = 50
d = 2

rng = MersenneTwister(1)
lambdas =  [10,20,40]
proba = [1/3,1/3,1/3]
points, labels = sample_poisson(rng, n - n_outliers, d, lambdas, proba)

outliers = sample_outliers(rng, n_outliers, d; scale = 120)
x = hcat(points, outliers)
labels_true = vcat(labels, zeros(Int, n_outliers))

scatter( x[1,:], x[2,:], c = labels_true, palette = :rainbow)

Partitionnement des données sur un exemple

Pour partitionner les données, nous utiliserons les paramètres suivants.

k = 3
α = 0.1
maxiter = 50
nstart = 50

50

Application de l'algorithme classique de $k$-means élagué

J.A. Cuesta-Albertos, A. Gordaliza, C. Matràn (1997)

Dans un premier temps, nous utilisons notre algorithme trimmed_bregman_clustering avec le carré de la norme Euclidienne euclidean.

tb_kmeans = trimmed_bregman_clustering( rng, x, k, α, euclidean, maxiter, nstart )
println("k-means : $(tb_kmeans.centers)")

k-means : [19.43082958404221 40.537339677644056 10.21567576495661; 20.266024052051026 39.91033785394563 9.795288930778936]

On observe trois groupes de même diamètre. Ainsi, de nombreuses données aberrantes sont associées au groupe des points générés selon la loi de Poisson de paramètre $(10,10)$. Ce groupe était sensé avoir un diamètre plus faible que les groupes de points issus des lois de Poisson de paramètres $(20,20)$ et $(40,40)$.

plot(tb_kmeans)

Choix de la divergence de Bregman associée à la loi de Poisson

Lorsque l'on utilise la divergence de Bregman associée à la loi de Poisson, les groupes sont de diamètres variables et sont particulièrement adaptés aux données. En particulier, les estimations tB_Poisson$centers des moyennes par les centres sont bien meilleures.

tb_poisson = trimmed_bregman_clustering( rng, x, k, α, poisson, maxiter, nstart )
println("poisson : $(tb_poisson.centers)")

poisson : [21.723795923078438 10.917323245970008 39.34836718371689; 20.137688511998164 11.120144074093618 39.974073758333596]

plot(tb_poisson)

Comparaison des performances

Nous mesurons directement la performance des deux partitionnements (avec le carré de la norme Euclidienne, et avec la divergence de Bregman associée à la loi de Poisson), à l'aide de l'information mutuelle normalisée.

println("k-means : $(mutualinfo( tb_kmeans.cluster, labels_true, true ))")
println("poisson : $(mutualinfo( tb_poisson.cluster, labels_true, true ))")

k-means : 0.6274362343840265
poisson : 0.6101795934423291

L'information mutuelle normalisée est supérieure pour la divergence de Bregman associée à la loi de Poisson. Ceci illustre le fait que sur cet exemple, l'utilisation de la bonne divergence permet d'améliorer le partitionnement, par rapport à un $k$-means élagué basique.

Mesure de la performance

Afin de s'assurer que la méthode avec la bonne divergence de Bregman est la plus performante, nous répétons l'expérience précédente replications_nb fois.

Pour ce faire, nous appliquons l'algorithme trimmed_bregman_clustering, sur replications_nb échantillons de taille $n = 1000$, sur des données générées selon la même procédure que l'exemple précédent.

La fonction performance.measurement permet de le faire.

sample_generator = (rng, n) -> sample_poisson(rng, n, d, lambdas, proba)
outliers_generator = (rng, n) -> sample_outliers(rng, n, d; scale = 120)

nmi_kmeans, _, _ = performance(n, n_outliers, k, α, sample_generator, outliers_generator, euclidean)
nmi_poisson, _, _ = performance(n, n_outliers, k, α, sample_generator, outliers_generator, poisson)

([0.5410506444342006, 0.5978382061714806, 0.6679595914345331, 0.5981197308183633, 0.6356421743201075, 0.5767036123293813, 0.5630512997896977, 0.5066321854022503, 0.671162908829301, 0.7308336595703248  …  0.43183799248823, 0.5398871861761185, 0.6536222204211154, 0.56412054154909, 0.5267476538001453, 0.6091343168740072, 0.7032707479649363, 0.677216074761488, 0.6289042453704504, 0.6209840411507159], 0.5863282353059076, 0.011734924654306193)

Les boîtes à moustaches permettent de se faire une idée de la répartition des NMI pour les deux méthodes différentes. On voit que la méthode utilisant la divergence de Bregman associée à la loi de Poisson est la plus performante.

using StatsPlots

boxplot( ones(100), nmi_kmeans, label = "kmeans" )
boxplot!( fill(2, 100), nmi_poisson, label = "poisson" )

Sélection des paramètres $k$ et $\alpha$

On garde le même jeu de données x.

vect_k = collect(1:5)
vect_α = sort([((0:2)./50)...,((1:4)./5)...])

rng = MersenneTwister(42)
nstart = 5

params_risks = select_parameters_nonincreasing(rng, vect_k, vect_α, x, poisson, maxiter, nstart)

plot(; title = "select parameters")
for (i,k) in enumerate(vect_k)
   plot!( vect_α, params_risks[i, :], label ="k=$k", markershape = :circle )
end
xlabel!("alpha")
ylabel!("NMI")

D'après la courbe, on voit qu'on gagne beaucoup à passer de 1 à 2 groupes, puis à passer de 2 à 3 groupes. Par contre, on gagne très peu, en termes de risque, à passer de 3 à 4 groupes ou à passer de 4 ou 5 groupes, car les courbes associées aux paramètres $k = 3$, $k = 4$ et $k = 5$ sont très proches. Ainsi, on choisit de partitionner les données en $k = 3$ groupes.

La courbe associée au paramètre $k = 3$ diminue fortement puis à une pente qui se stabilise aux alentours de $\alpha = 0.04$.

Pour plus de précisions concernant le choix du paramètre $\alpha$, nous pouvons nous concentrer que la courbe $k = 3$ en augmentant la valeur de nstart et en nous concentrant sur les petites valeurs de $\alpha$.

vec_k = [3]
vec_α = collect(0:15) ./ 200
params_risks = select_parameters_nonincreasing(rng, vec_k, vec_α, x, poisson, maxiter, nstart)

plot(vec_α, params_risks[1, :], markershape = :circle)

On ne voit pas de changement radical de pente mais on voit que la pente se stabilise après $\alpha = 0.04$. Nous choisissons le paramètre $\alpha = 0.04$.

k, α = 3, 0.04
tb = trimmed_bregman_clustering( rng, x, k, α, poisson, maxiter, nstart )
plot(tb)

Application au partitionnement de textes d'auteurs

Les données des textes d'auteurs sont enregistrées dans la variable data. Les commandes utilisées pour l'affichage étaient les suivantes.

using DataFrames
using DelimitedFiles
using GeometricClusterAnalysis
using NamedArrays
using Random

rng = MersenneTwister(2022)

table = readdlm(joinpath("assets","textes.txt"))

df = DataFrame( hcat(table[2:end,1], table[2:end,2:end]), vec(vcat("authors",table[1,1:end-1])), makeunique=true)

dft = DataFrame([[names(df)[2:end]]; collect.(eachrow(df[:,2:end]))], [:column; Symbol.(axes(df, 1))])
rename!(dft, String.(vcat("authors",values(df[:,1]))))

data = NamedArray( table[2:end,2:end]', (names(df)[2:end], df.authors ), ("Rows", "Cols"))

authors = ["God", "Doyle", "Dickens", "Hawthorne",  "Obama", "Twain"]
authors_names = ["Bible",  "Conan Doyle", "Dickens", "Hawthorne", "Obama", "Twain"]
true_labels = [sum(count.(author, names(df))) for author in authors]

Afin de pouvoir représenter les données, nous utiliserons la fonction suivante.

R"""
plot_clustering <- function(axis1 = 1, axis2 = 2, labels, title = "Textes d'auteurs - Partitionnement"){
  to_plot = data.frame(lda = lda$li, Etiquettes =  as.factor(labels), authors_names = as.factor(authors_names))
  ggplot(to_plot, aes(x = lda$li[,axis1], y =lda$li[,axis2],col = Etiquettes, shape = authors_names))+ xlab(paste("Axe ",axis1)) + ylab(paste("Axe ",axis2))+ 
  scale_shape_discrete(name="Auteur") + labs (title = title) + geom_point()}
"""

Partitionnement des données

Pour partitionner les données, nous utiliserons les paramètres suivants.

k = 4
alpha = 20/209 # La vraie proportion de donnees aberrantes vaut : 20/209 car il y a 15+5 textes issus de la bible et du discours de Obama.
maxiter = 50
nstart = 50

Application de l'algorithme classique de $k$-means élagué [@Cuesta-Albertos1997]

tb_authors_kmeans = trimmed_bregman_clustering(rng, data, k, alpha, euclidean, maxiter, nstart)

#plot_clustering(1,2,tB_authors_kmeans$cluster)
#plot_clustering(3,4,tB_authors_kmeans$cluster)

Choix de la divergence de Bregman associée à la loi de Poisson

tb_authors_poisson = trimmed_bregman_clustering(rng, data, k, alpha, poisson, maxiter, nstart)

#plot_clustering(1,2,tB_authors_Poisson$cluster)
#plot_clustering(3,4,tB_authors_Poisson$cluster)
"""

En utilisant la divergence de Bregman associée à la loi de Poisson, nous voyons que notre méthode de partitionnement fonctionne très bien avec les paramètres k = 4 et alpha = 20/209. En effet, les données aberrantes sont bien les textes de Obama et de la bible. Par ailleurs, les autres textes sont plutôt bien partitionnés.

Comparaison des performances

Nous mesurons directement la performance des deux partitionnements (avec le carré de la norme Euclidienne, et avec la divergence de Bregman associée à la loi de Poisson), à l'aide de l'information mutuelle normalisée.

Vraies etiquettes ou les textes issus de la bible et du discours de Obama ont la meme etiquette :

R"true_labels[true_labels == 5] = 1"

Pour le k-means elague :

R"""
NMI(true_labels,tB_authors_kmeans$cluster, variant="sqrt")
"""

Pour le partitionnement elague avec divergence de Bregman associee a la loi de Poisson :

R"""
NMI(true_labels,tB_authors_Poisson$cluster, variant="sqrt")
"""

L'information mutuelle normalisée est bien supérieure pour la divergence de Bregman associée à la loi de Poisson. Ceci illustre le fait que l'utilisation de la bonne divergence permet d'améliorer le partitionnement, par rapport à un $k$-means élagué basique. En effet, le nombre d'apparitions d'un mot dans un texte d'une longueur donnée, écrit par un même auteur, peut-être modélisé par une variable aléatoire de loi de Poisson. L'indépendance entre les nombres d'apparition des mots n'est pas forcément réaliste, mais on ne tient compte que d'une certaine proportion des mots (les 50 les plus présents). On peut donc faire cette approximation. On pourra utiliser la divergence associée à la loi de Poisson.

Sélection des paramètres $k$ et $\alpha$

Affichons maintenant les courbes de risque en fonction de $k$ et de $\alpha$ pour voir si d'autres choix de paramètres auraient été judicieux. En pratique, c'est important de réaliser cette étape, car nous ne sommes pas sensés connaître le jeu de données, ni le nombre de données aberrantes.

R"""

vect_k = 1:6
vect_alpha = c((1:5)/50,0.15,0.25,0.75,0.85,0.9)
nstart = 20
set.seed(1)
params_risks = select.parameters(vect_k,vect_alpha,data,divergence_Poisson,maxiter,nstart,.export = c('divergence_Poisson','divergence_Poisson','data','nstart','maxiter'),force_nonincreasing = TRUE)

params_risks$k = as.factor(params_risks$k)
ggplot(params_risks, aes(x = alpha, y = risk, group = k, color = k))+   geom_line() +   geom_point()
"""

Pour sélectionner les paramètres k et alpha, on va se concentrer sur différents segments de valeurs de alpha. Pour alpha supérieur à 0.15, on voit qu'on gagne beaucoup à passer de 1 à 2 groupes, puis à passer de 2 à 3 groupes. On choisirait donc k = 3 et alphade l'ordre de $0.15$ correspondant au changement de pente de la courbe k = 3.

Pour alpha inférieur à 0.15, on voit qu'on gagne beaucoup à passer de 1 à 2 groupes, à passer de 2 à 3 groupes, puis à passer de 3 à 4 groupes. Par contre, on gagne très peu, en termes de risque, à passer de 4 à 5 groupes ou à passer de 5 ou 6 groupes, car les courbes associées aux paramètres $k = 4$, $k = 5$ et $k = 6$ sont très proches. Ainsi, on choisit de partitionner les données en $k = 4$ groupes.

La courbe associée au paramètre $k = 4$ diminue fortement puis a une pente qui se stabilise aux alentours de $\alpha = 0.1$.

Enfin, puisqu'il y a un saut avant la courbe $k = 6$, nous pouvons aussi choisir le paramètre k = 6, auquel cas alpha = 0, nous ne considérons aucune donnée aberrante.

Remarquons que le fait que notre méthode soit initialisée avec des centres aléatoires implique que les courbes représentant le risque en fonction des paramètres $k$ et $\alpha$ puissent varier, assez fortement, d'une fois à l'autre. En particulier, le commentaire, ne correspond peut-être pas complètement à la figure représentée. Pour plus de robustesse, il aurait fallu augmenter la valeur de nstart et donc aussi le temps d'exécution. Ces courbes pour sélectionner les paramètres k et alpha sont donc surtout indicatives.

Finalement, voici les trois partitionnements obtenus à l'aide des 3 choix de paires de paramètres.

R"""
tB = Trimmed_Bregman_clustering(data,3,0.15,divergence_Poisson,maxiter = 50, nstart = 50)
plot_clustering(1,2,tB$cluster)
"""
# -

Les textes de Twain, de la bible et du discours de Obama sont considérées comme des données aberrantes.

R"""
tB = Trimmed_Bregman_clustering(data,4,0.1,divergence_Poisson,maxiter = 50, nstart = 50)
plot_clustering(1,2,tB$cluster)
"""
# -

Les textes de la bible et du discours de Obama sont considérés comme des données aberrantes.

R"""
tB = Trimmed_Bregman_clustering(data,6,0,divergence_Poisson,maxiter = 50, nstart = 50)
plot_clustering(1,2,tB$cluster)
"""
# -

On obtient 6 groupes correspondant aux textes des 4 auteurs différents, aux textes de la bible et au discours de Obama.