Données de loi de Poisson en dimension 1

Simulation des variables selon un mélange de lois de Poisson

La fonction sample_poisson permet de simuler des variables aléatoires selon un mélange de $k$ lois de Poisson en dimension $d$, de paramètres donnés par la matrice lambdas de taille $k\times d$. Les probabilités associées à chaque composante du mélange sont données dans le vecteur proba.

La fonction sample_outliers permet de simuler des variables aléatoires uniformément sur l'hypercube $[0,L]^d$. On utilisera cette fonction pour générer des données aberrantes.

GeometricClusterAnalysis.sample_poisson — Function

sample_poisson(rng, n, d, lambdas, proba)

GeometricClusterAnalysis.sample_outliers — Function

sample_outliers(rng, n_outliers, d; scale = 1)

On génère un premier échantillon de 950 points de loi de Poisson de paramètre $10$, $20$ ou $40$ avec probabilité $\frac13$, puis un échantillon de 50 données aberrantes de loi uniforme sur $[0,120]$. On note x l'échantillon ainsi obtenu.

using GeometricClusterAnalysis
import GeometricClusterAnalysis: sample_poisson, sample_outliers, performance
using Plots
using Random

n = 1000
n_outliers = 50
d = 1

rng = MersenneTwister(1)
lambdas =  [10,20,40]
proba = [1/3,1/3,1/3]
points, labels = sample_poisson(rng, n - n_outliers, d, lambdas, proba)

outliers = sample_outliers(rng, n_outliers, 1; scale = 120)

x = hcat(points, outliers)
labels_true = vcat(labels, zeros(Int, n_outliers))
scatter( x[1,:], c = labels_true, palette = :rainbow)

Partitionnement des données sur un exemple

Pour partitionner les données, nous utiliserons les paramètres suivants.

k = 3 # Nombre de groupes dans le partitionnement
alpha = 0.04 # Proportion de donnees aberrantes
maxiter = 50 # Nombre maximal d'iterations
nstart = 20 # Nombre de departs

20

Application de l'algorithme classique de $k$-means élagué J.A. Cuesta-Albertos, A. Gordaliza, C. Matràn (1997)

Dans un premier temps, nous utilisons notre algorithme trimmed_bregman_clustering avec le carré de la norme Euclidienne euclidean.

tb_kmeans = trimmed_bregman_clustering(rng, x, k, alpha, euclidean, maxiter, nstart)
tb_kmeans.centers

1×3 Matrix{Float64}:
 22.116  40.9987  10.3412

Nous avons effectué un simple algorithme de $k$-means élagué, comme J.A. Cuesta-Albertos, A. Gordaliza, C. Matràn (1997). On voit trois groupes de même diamètre. Ce qui fait que le groupe centré en $10$ contient aussi des points du groupe centré en $20$. En particulier, les estimations tB_kmeans.centers des moyennes par les centres ne sont pas très bonnes. Les moyennes sont supérieures aux vraies moyennes $10$, $20$ et $40$.

plot(tb_kmeans)

Choix de la divergence de Bregman associée à la loi de Poisson

Lorsque l'on utilise la divergence de Bregman associée à la loi de Poisson, les groupes sont de diamètres variables et sont particulièrement adaptés aux données. En particulier, les estimations tB_Poisson$centers des moyennes par les centres sont bien meilleures.

tb_poisson = trimmed_bregman_clustering(rng, x, k, alpha, poisson, maxiter, nstart)
tb_poisson.centers

1×3 Matrix{Float64}:
 40.7604  20.6136  9.71071

plot(tb_poisson)

Comparaison des performances

Nous mesurons directement la performance des deux partitionnements (avec le carré de la norme Euclidienne, et avec la divergence de Bregman associée à la loi de Poisson), à l'aide de l'information mutuelle normalisée.

Pour le k-means elague :

import Clustering: mutualinfo

println(mutualinfo(labels_true,tb_kmeans.cluster, normed = true))

0.641631775173803

Pour le partitionnement elague avec divergence de Bregman associee a la loi de Poisson :

println(mutualinfo(labels_true,tb_poisson.cluster, normed = true))

0.6826165935516633

L'information mutuelle normalisée est supérieure pour la divergence de Bregman associée à la loi de Poisson. Ceci illustre le fait que sur cet exemple, l'utilisation de la bonne divergence permet d'améliorer le partitionnement, par rapport à un $k$-means élagué basique.

Mesure de la performance

Afin de s'assurer que la méthode avec la bonne divergence de Bregman est la plus performante, nous répétons l'expérience précédente replications fois.

Pour ce faire, nous appliquons l'algorithme trimmed_bregman_clustering, sur replications échantillons de taille $n = 1000$, sur des données générées selon la même procédure que l'exemple précédent.

La fonction performance permet de le faire.

sample_generator = (rng, n) -> sample_poisson(rng, n, d, lambdas, proba)
outliers_generator = (rng, n) -> sample_outliers(rng, n, d; scale = 120)

#3 (generic function with 1 method)

Valeurs par défault: maxiter = 100, nstart = 10, replications = 100

n = 1200
n_outliers = 200
k = 3
alpha = 0.1
nmi_kmeans, _, _ = performance(n, n_outliers, k, alpha, sample_generator, outliers_generator, euclidean)
nmi_poisson, _, _ = performance(n, n_outliers, k, alpha, sample_generator, outliers_generator, poisson)

([0.5839398784033235, 0.6202318416861812, 0.6101178140792, 0.6374897347566113, 0.6540671777145892, 0.6083343936925494, 0.6304838453433942, 0.6306156841785864, 0.6336810374826111, 0.6527637354246719  …  0.6248676870772997, 0.6257423603470843, 0.6304698035282263, 0.6376606605106416, 0.6270816412278291, 0.6273014539714742, 0.6195508045521951, 0.6298634762917615, 0.6471578693973343, 0.6209459002804606], 0.6300521037537219, 0.003665377094835531)

Les boîtes à moustaches permettent de se faire une idée de la répartition des NMI pour les deux méthodes différentes. On voit que la méthode utilisant la divergence de Bregman associée à la loi de Poisson est la plus performante.

using StatsPlots

boxplot( ones(100), nmi_kmeans, label = "kmeans" )
boxplot!( fill(2, 100), nmi_poisson, label = "poisson" )

Sélection des paramètres $k$ et $\alpha$

On garde le même jeu de données x.

vect_k = collect(1:5)
vect_alpha = sort([((0:2)./50)...,((1:4)./5)...])

params_risks = select_parameters_nonincreasing(rng, vect_k, vect_alpha, x, poisson, maxiter, nstart)

plot(; title = "select parameters")
for (i,k) in enumerate(vect_k)
   plot!( vect_alpha, params_risks[i, :], label ="k=$k", markershape = :circle )
end
xlabel!("alpha")
ylabel!("NMI")

D'après la courbe, on voit qu'on gagne beaucoup à passer de 1 à 2 groupes, puis à passer de 2 à 3 groupes. Par contre, on gagne très peu, en termes de risque, à passer de 3 à 4 groupes ou à passer de 4 à 5 groupes, car les courbes associées aux paramètres $k = 3$, $k = 4$ et $k = 5$ sont très proches. Ainsi, on choisit de partitionner les données en $k = 3$ groupes.

La courbe associée au paramètre $k = 3$ diminue fortement puis à une pente qui se stabilise aux alentours de $\alpha = 0.04$.

Pour plus de précisions concernant le choix du paramètre $\alpha$, nous pouvons nous concentrer sur la courbe $k = 3$ en augmentant la valeur de nstart et en nous concentrant sur les petites valeurs de $\alpha$.

vect_k = [3]
vec_alpha = collect(0:15) ./ 200
params_risks = select_parameters_nonincreasing(rng, [3], vec_alpha, x, poisson, maxiter, 5)

plot(vec_alpha, params_risks[1, :], markershape = :circle)

On ne voit pas de changement radical de pente mais on voit que la pente se stabilise après $\alpha = 0.03$. Nous choisissons le paramètre $\alpha = 0.03$.

Voici finalement le partitionnement obtenu après sélection des paramètres k et alpha selon l'heuristique.

k, alpha = 3, 0.03
tb_poisson = trimmed_bregman_clustering( rng, x, k, alpha, poisson, maxiter, nstart )
tb_poisson.centers

1×3 Matrix{Float64}:
 41.0419  20.6136  9.57292

plot( tb_poisson )